requirements.txt
Django==5.1
django-restframework==0.0.1
djangorestframework==3.15.2
flake8==7.1.1
isort==5.13.2
pytest==8.3.2
psycopg2-binary==2.9.9
djangorestframework-simplejwt==5.3.1
pillow==10.4.0